Create Question2.py
Bubbles
There has been a global pandemic. People must stay within their bubbles. If two people have physical contact, they are part of the same bubble.

Write a function bubbles(physical_contact_info) that takes physical contact information about a group of people and determines the bubbles.

Input
The only argument passed to the function is the textual representation of an undirected graph. Each vertex corresponds to a person. If (and only if) two people have physical contact, there is an edge between the corresponding vertices.

Output
The function must return a list. The order of elements in the list does not matter. The number of elements in the list must be equal to the number of bubbles. Each element of the list is a set of people who are part of the same bubble (use a Python set or a list but the order does not matter).