Create Question3.py
Package management
Computer programs can depend on each other. Sometimes a program must be built (or installed) before another program that depends on it can be built (or installed).

Write a function build_order(dependencies) that takes a description of the dependencies between a number of programs and returns a valid order for the build process. You can assume that there will always be a solution.

Input
The input is a string representation of a directed graph. The number of vertices, n, is the number of programs of interest. When there is a directed edge from one program to another, the former should be built before the latter.

Output
The function must return a list of length n of programs in a valid order build process (from left to right). If there are multiple solutions, any one of them is acceptable.