Create Question4.py
Snow
A heavy snow has caused road closures in the city. The first priority is to make sure all locations in the city are safely reachable and connected to each other so that essential services can be provided. The goal is to clear the least amount of road surface so that all locations are reachable.

Write a function which_segments(city_map) that takes the map of the city and returns a list of road segments that must be cleared so that there is a clear path between any two locations and the total length of the cleaned-up road segments is minimised.

Input
The city map is given as the textual representation of an undirected weighted graph. Each vertex in the graph corresponds to one location in the city. Each (two-way) road segment is represented as an edge which connects two locations together and the weight of the edge is the length of road segment. Additionally, the following assumptions are made:

the given city has at least one location;
there is a path between any two locations;
Output
The output is a list of road segments that must be cleared. Each segment is a tuple (pair) of two location numbers. The smaller number should appear first. If the solution is not unique, it does not matter which solution is returned by your function.