Create Question5.py
Batteries
A distribution company is upgrading its delivery fleet to electric vehicles. In doing so they have to determine the battery capacity required for vehicles. The following factors must be considered:

The company has a single depot and it delivers to any location in the city that is reachable from the depot.
A delivery vehicle can travel two units of distance for every 3 units of battery capacity (that is charged).
It is assumed that given a destination, the vehicle will always take a path with the shortest distance in order to go to the destination and come back.
A vehicle with a fully-charged battery must be able to make a trip from the depot to any one location in the city and return to the depot with at least 25% charge left in the battery.
Write a function min_capacity(city_map, depot_position) that takes the map of a city (described in terms of locations, segments of road in between locations, and the length of each segment), and the position of the depot, and returns the minimum battery capacity required to meet the above-mentioned criteria.

Input
The city_map is the textual representation of an undirected weighted graph. There is a vertex for each location in the city. There is an edge for each segment of road. All roads are two-way. The length of the road segment is the weight of the edge.

The parameter depot_position is a location in the city where the depot is located.

Output
The function must return an integer that is the minimum capacity required for the battery.